pyproject.toml
[project]
name = "fbref-chatbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "blinker==1.9.0",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cloudscraper>=1.2.71",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "duckdb>=1.3.2",
    "duckdb-engine>=0.17.0",
    "flask==3.1.1",
    "flask-cors>=6.0.1",
    "frozenlist==1.7.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.72",
    "langchain-experimental==0.3.4",
    "langchain-google-genai>=2.1.9",
    "langchain-text-splitters==0.3.9",
    "langsmith==0.4.9",
    "lxml>=6.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "multidict==6.6.3",
    "mypy-extensions==1.1.0",
    "numpy==2.3.2",
    "openai==1.98.0",
    "orjson==3.11.1",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "pandas==2.3.1",
    "propcache==0.3.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "selenium==4.34.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.7",
    "sqlalchemy==2.0.42",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "trio==0.30.0",
    "trio-websocket==0.12.2",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "undetected-chromedriver>=3.5.5",
    "urllib3==2.5.0",
    "websocket-client==1.8.0",
    "werkzeug==3.1.3",
    "wsproto==1.2.0",
    "yarl==1.20.1",
    "zstandard==0.23.0",
]
